+++ addon/views/ivy-sortable.js
@@ -6,8 +6,6 @@ export default Ember.CollectionView.extend(Ember.TargetActionSupport, {
   tagName: 'ul',
 
   destroySortable: Ember.on('willDestroyElement', function() {
-    this.removeObserver('disabled', this, this._disabledDidChange);
-
     this.$().sortable('destroy');
   }),
 
@@ -20,8 +18,12 @@ export default Ember.CollectionView.extend(Ember.TargetActionSupport, {
 
     this.$().sortable(opts);
 
-    this.addObserver('disabled', this, this._disabledDidChange);
-    this._disabledDidChange();
+    Ember.EnumerableUtils.forEach([
+      'axis', 'containment', 'cursor', 'cursorAt', 'delay', 'disabled',
+      'distance', 'forceHelperSize', 'forcePlaceholderSize', 'grid', 'helper',
+      'opacity', 'placeholder', 'revert', 'scroll', 'scrollSensitivity',
+      'scrollSpeed', 'tolerance', 'zIndex'
+    ], this._bindSortableOption, this);
   }),
 
   move: function(oldIndex, newIndex) {
@@ -87,6 +89,18 @@ export default Ember.CollectionView.extend(Ember.TargetActionSupport, {
     this.move(oldIndex, newIndex);
   },
 
+  _bindSortableOption: function(key) {
+    this.addObserver(key, this, this._optionDidChange);
+
+    if (key in this) {
+      this._optionDidChange(this, key);
+    }
+
+    this.on('willDestroyElement', this, function() {
+      this.removeObserver(key, this, this._optionDidChange);
+    });
+  },
+
   _disableArrayObservers: function(content, callback) {
     content.removeArrayObserver(this);
     try {
@@ -96,7 +110,7 @@ export default Ember.CollectionView.extend(Ember.TargetActionSupport, {
     }
   },
 
-  _disabledDidChange: function() {
-    this.$().sortable(this.get('disabled') ? 'disable' : 'enable');
+  _optionDidChange: function(sender, key) {
+    this.$().sortable('option', key, this.get(key));
   }
 });