+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## master
 
+* Add jQuery UI Sortable option bindings.
+
 ## 0.2.0
 
 * Convert to an ember-cli addon.